Seperated Build Process from Cascade
Created new class HT::Builder which when run takes a cascade, the input data and the layer to build to.
Although there is nothing prevented it, the design is to use one builder per thread or application.

Blocks passed when creating the cascade are now run in the context of the builder instance. The methods
provided for convenience in the block are now called on the builder instance instead of the cascade instance.

Added some convenenience methods to the cascade and some tests.