cascade names now have a purpose
If a name is passed into HT::Cascade.new the string value of the argument
is used as a key to access the cascade again. This allows code to access the cascade
without reference to the object and only knowledge of its key.

To access the cascade use HT::Cascade.[] which takes either the string key used
to register the cascade or the original argument passed to initialize (which will be converted to its
string value for you).

You can flush the existing list of global cascades with HT::Cascade.flush_global

This is the first step to allowing cascades to be reopened for modification.